pyproject.toml
[project]
name = "tt-agent"
version = "0.0.1"
description = "dear leaders favourite agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
